backend/requirements.txt
Django==3.2.*
Pillow==9.4.0
django_filter==22.1
djangorestframework-simplejwt==4.8.0
djangorestframework==3.14.0
djoser==2.1.0
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-return==1.2.0
flake8==5.0.4
gunicorn==20.1.0
isort==5.11.4
pep8-naming==0.13.3
psycopg2-binary==2.9.5
python-dotenv==0.21.1
drf-extra-fields==3.4.0